refactor(vite): reuse logics and improve code splitting
